chore(ui): remove unused Combobox component and its related files to clean up the codebase fix(package.json): remove unused @radix-ui/react-select dependency to reduce bundle size
chore(ui): remove Combobox component and related files to streamline the codebase
fix(input): adjust input styles for better consistency and appearance
fix(select): update Select component exports to simplify usage in stories
feat(select.stories): enhance Select stories with new options and placeholders for better demonstration

test(select): enhance tests for Select component with additional cases and improved structure to ensure comprehensive coverage and better maintainability

refactor(select.tsx): simplify Select component structure and improve styling consistency

This change refactors the Select component by removing unnecessary
wrappers and imports, enhancing readability and maintainability.
It also updates the styling to ensure a consistent design across
different sizes of the select component.

feat(select): refactor Select component to support combobox functionality and improve accessibility

This change introduces a new combobox feature to the Select component, allowing users to search through options. It enhances the user experience by providing a more interactive and accessible way to select options. The refactor also improves the structure and readability of the code.

feat(index.css): add new color variable --color-neutral-950 for enhanced styling options
refactor(registry.json): update dependencies for Select component and remove Combobox component to streamline the registry
