--- packages/ui/src/components/ui/combobox/index.tsx
@@ -1,1 +1,0 @@
-export { Combobox } from './combobox'
