+++ packages/ui/src/components/ui/select/index.tsx
@@ -1,12 +1,1 @@
-export {
-  Select,
-  SelectContent,
-  SelectGroup,
-  SelectItem,
-  SelectLabel,
-  SelectScrollDownButton,
-  SelectScrollUpButton,
-  SelectTrigger,
-  SelectValue,
-  SelectInput,
-} from './select'
+export { Select } from './select'
