+++ packages/ui/src/components/ui/select/select.tsx
@@ -1,24 +1,13 @@
-import * as React from 'react'
-import * as SelectPrimitive from '@radix-ui/react-select'
-import { CheckIcon, ChevronDownIcon, ChevronUpIcon } from 'lucide-react'
+import React, { useState, useRef, useEffect } from 'react'
+import * as Popover from '@radix-ui/react-popover'
+import { Search, Check, ChevronsUpDown } from 'lucide-react'
 import { cva, type VariantProps } from 'class-variance-authority'
+
 import { cn } from '@/lib/utils'
-
-function Select({ ...props }: React.ComponentProps<typeof SelectPrimitive.Root>) {
-  return <SelectPrimitive.Root data-slot="select" {...props} />
-}
-
-function SelectGroup({ ...props }: React.ComponentProps<typeof SelectPrimitive.Group>) {
-  return <SelectPrimitive.Group data-slot="select-group" {...props} />
-}
-
-function SelectValue({ ...props }: React.ComponentProps<typeof SelectPrimitive.Value>) {
-  return <SelectPrimitive.Value data-slot="select-value" {...props} />
-}
 
 const selectVariants = cva(
   `
-    text-white data-[placeholder]:text-muted-foreground [&_svg:not([class*='text-'])]:text-muted-foreground focus-visible:border-ring focus-visible:ring-ring/50 dark:bg-input/30 dark:hover:bg-input/50 flex w-fit items-center justify-between gap-2 rounded-md  bg-transparent px-3 py-2 text-sm whitespace-nowrap transition-[color,box-shadow] outline-none focus-visible:ring-[3px]  data-[size=default]:h-9 data-[size=sm]:h-8 *:data-[slot=select-value]:line-clamp-1 *:data-[slot=select-value]:flex *:data-[slot=select-value]:items-center *:data-[slot=select-value]:gap-2 [&_svg]:pointer-events-none [&_svg]:shrink-0 [&_svg:not([class*='size-'])]:size-4
+    text-white data-[placeholder]:text-muted-foreground [&_svg:not([class*='text-'])]:text-muted-foreground focus-visible:border-ring focus-visible:ring-ring/50  flex w-fit items-center justify-between gap-2 rounded-[0.25rem]  bg-neutral-900 px-3 py-2 text-sm whitespace-nowrap transition-[color,box-shadow] outline-none focus-visible:ring-[3px]  data-[size=default]:h-9 data-[size=sm]:h-8 *:data-[slot=select-value]:line-clamp-1 *:data-[slot=select-value]:flex *:data-[slot=select-value]:items-center *:data-[slot=select-value]:gap-2 [&_svg]:pointer-events-none [&_svg]:shrink-0 [&_svg:not([class*='size-'])]:size-4
 
     shadow-[0_1px_2px_rgba(0,0,0,0.05)]  border border-[var(--secondary-border)]
 
@@ -33,10 +22,10 @@ const selectVariants = cva(
   {
     variants: {
       size: {
-        xs: 'px-[8px] h-[26px] text-xs',
-        sm: 'px-[8px] h-[32px] text-sm',
-        md: 'px-[8px] h-[36px] text-base',
-        lg: 'px-[8px] h-[40px] text-base',
+        xs: 'px-[0.5rem] h-[1.5rem] text-xs',
+        sm: 'px-[0.5rem] h-[2rem] text-sm',
+        md: 'px-[0.5rem] h-[2.25rem] text-sm',
+        lg: 'px-[0.5rem] h-[2.5rem] text-sm',
       },
     },
     defaultVariants: {
@@ -45,176 +34,188 @@ const selectVariants = cva(
   },
 )
 
-function SelectTrigger({
-  className,
-  size,
-  children,
-  ...props
-}: React.ComponentProps<typeof SelectPrimitive.Trigger> & VariantProps<typeof selectVariants>) {
-  const isInvalid = props['aria-invalid']
-
-  return (
-    <SelectPrimitive.Trigger
-      data-slot="select-trigger"
-      className={cn(selectVariants({ size, className }), 'group')}
-      {...props}
-    >
-      {children}
-      <SelectPrimitive.Icon asChild>
-        <ChevronDownIcon
-          data-testid="select-chevron-icon"
-          className={`size-[15px] text-[var(--muted-foreground)]   ${!isInvalid && 'group-hover:text-[var(--primary)]'} `}
-        />
-      </SelectPrimitive.Icon>
-    </SelectPrimitive.Trigger>
-  )
-}
-
-function SelectContent({
-  className,
-  children,
-  position = 'popper',
-  ...props
-}: React.ComponentProps<typeof SelectPrimitive.Content>) {
-  return (
-    <SelectPrimitive.Portal>
-      <SelectPrimitive.Content
-        data-slot="select-content"
-        className={cn(
-          'bg-popover text-popover-foreground data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 relative z-50 max-h-(--radix-select-content-available-height) min-w-[8rem] origin-(--radix-select-content-transform-origin) overflow-x-hidden overflow-y-auto rounded-md border shadow-md',
-          position === 'popper' &&
-            'data-[side=bottom]:translate-y-1 data-[side=left]:-translate-x-1 data-[side=right]:translate-x-1 data-[side=top]:-translate-y-1',
-          className,
-        )}
-        position={position}
-        {...props}
-      >
-        <SelectScrollUpButton />
-        <SelectPrimitive.Viewport
-          className={cn(
-            'p-1',
-            position === 'popper' &&
-              'h-[var(--radix-select-trigger-height)] w-full min-w-[var(--radix-select-trigger-width)] scroll-my-1',
-          )}
-        >
-          {children}
-        </SelectPrimitive.Viewport>
-        <SelectScrollDownButton />
-      </SelectPrimitive.Content>
-    </SelectPrimitive.Portal>
-  )
-}
-
-function SelectLabel({ className, ...props }: React.ComponentProps<typeof SelectPrimitive.Label>) {
-  return (
-    <SelectPrimitive.Label
-      data-slot="select-label"
-      className={cn('text-muted-foreground px-2 py-1.5 text-xs', className)}
-      {...props}
-    />
-  )
-}
-
-function SelectItem({
-  className,
-  children,
-  ...props
-}: React.ComponentProps<typeof SelectPrimitive.Item>) {
-  return (
-    <SelectPrimitive.Item
-      data-slot="select-item"
-      className={cn(
-        "focus:bg-accent focus:text-accent-foreground [&_svg:not([class*='text-'])]:text-muted-foreground relative flex w-full cursor-default items-center gap-2 rounded-sm py-1.5 pr-8 pl-2 text-sm outline-hidden select-none data-[disabled]:pointer-events-none data-[disabled]:opacity-50 [&_svg]:pointer-events-none [&_svg]:shrink-0 [&_svg:not([class*='size-'])]:size-4 *:[span]:last:flex *:[span]:last:items-center *:[span]:last:gap-2",
-        className,
-      )}
-      {...props}
-    >
-      <span className="absolute right-2 flex size-3.5 items-center justify-center">
-        <SelectPrimitive.ItemIndicator>
-          <CheckIcon className="size-4" />
-        </SelectPrimitive.ItemIndicator>
-      </span>
-      <SelectPrimitive.ItemText>{children}</SelectPrimitive.ItemText>
-    </SelectPrimitive.Item>
-  )
-}
-
-function SelectScrollUpButton({
-  className,
-  ...props
-}: React.ComponentProps<typeof SelectPrimitive.ScrollUpButton>) {
-  return (
-    <SelectPrimitive.ScrollUpButton
-      data-slot="select-scroll-up-button"
-      className={cn('flex cursor-default items-center justify-center py-1', className)}
-      {...props}
-    >
-      <ChevronUpIcon className="size-4" />
-    </SelectPrimitive.ScrollUpButton>
-  )
-}
-
-function SelectScrollDownButton({
-  className,
-  ...props
-}: React.ComponentProps<typeof SelectPrimitive.ScrollDownButton>) {
-  return (
-    <SelectPrimitive.ScrollDownButton
-      data-slot="select-scroll-down-button"
-      className={cn('flex cursor-default items-center justify-center py-1', className)}
-      {...props}
-    >
-      <ChevronDownIcon className="size-4" />
-    </SelectPrimitive.ScrollDownButton>
-  )
-}
-
-type Option = { label: string; value: string }
-
-interface SelectInputProps
-  extends Omit<React.ComponentProps<typeof SelectPrimitive.Root>, 'value' | 'onValueChange'>,
-    VariantProps<typeof selectVariants> {
-  placeholder: string
-  selectLabel?: string
-  options: Option[]
-  disabled?: boolean
-  className?: string
-  'aria-invalid'?: boolean
-}
-
-const SelectInput = React.forwardRef<React.ComponentRef<typeof SelectTrigger>, SelectInputProps>(
-  ({ options, size, className, placeholder, selectLabel, ...props }, ref) => {
-    return (
-      <Select>
-        <SelectTrigger {...props} ref={ref} className={className} size={size}>
-          <SelectValue placeholder={placeholder} />
-        </SelectTrigger>
-        <SelectContent>
-          <SelectGroup>
-            <SelectLabel>{selectLabel ?? placeholder}</SelectLabel>
-            {options.map((option) => (
-              <SelectItem key={option.value} value={option.value}>
-                {option.label}
-              </SelectItem>
-            ))}
-          </SelectGroup>
-        </SelectContent>
-      </Select>
-    )
+const popoverContentItemVariants = cva(
+  `w-full flex items-center gap-2 cursor-pointer outline-none hover:bg-neutral-800 text-white`,
+  {
+    variants: {
+      size: {
+        xs: 'px-2 py-1 text-xs',
+        sm: 'px-2 py-1.5 text-sm',
+        md: 'px-2 py-2 text-sm',
+        lg: 'px-2 py-2.5 text-sm',
+      },
+    },
+    defaultVariants: {
+      size: 'md',
+    },
   },
 )
 
-SelectInput.displayName = 'SelectInput'
+type SelectOption = {
+  value: string
+  label: string
+}
 
-export {
-  Select,
-  SelectContent,
-  SelectGroup,
-  SelectItem,
-  SelectLabel,
-  SelectScrollDownButton,
-  SelectScrollUpButton,
-  SelectTrigger,
-  SelectValue,
-  SelectInput,
+type SelectProps = {
+  options: SelectOption[]
+  defaultValue?: string
+  onChange: (value: string) => void
+  placeholder?: string
+  searchPlaceholder?: string
+  emptyStatePlaceholder?: string
+  className?: string
+  isCombobox?: boolean
+  disabled?: boolean
+  'aria-invalid'?: boolean
+} & VariantProps<typeof selectVariants>
+
+export const Select: React.FC<SelectProps> = ({
+  options,
+  defaultValue = '',
+  onChange,
+  placeholder = 'Selecione',
+  searchPlaceholder = 'Pesquisar...',
+  emptyStatePlaceholder = 'Sem dados',
+  className,
+  isCombobox,
+  size,
+  disabled,
+  ...rest
+}) => {
+  const [search, setSearch] = useState('')
+  const [internalValue, setInternalValue] = useState('')
+  const [open, setOpen] = useState(false)
+  const [focusedIndex, setFocusedIndex] = useState<number>(-1)
+  const itemRefs = useRef<HTMLButtonElement[]>([])
+  const inputRef = useRef<HTMLInputElement>(null)
+
+  const isInvalid = rest['aria-invalid']
+
+  const filteredOptions = options.filter((opt) =>
+    opt.label.toLowerCase().includes(search.toLowerCase()),
+  )
+
+  function handleSelect(optionValue: string) {
+    const newValue = optionValue === internalValue ? '' : optionValue
+    setInternalValue(newValue)
+    onChange(newValue)
+    setOpen(false)
+  }
+
+  function clearSearch() {
+    setSearch('')
+  }
+
+  function handleSearch(e: React.ChangeEvent<HTMLInputElement>) {
+    setSearch(e.target.value)
+    setFocusedIndex(0)
+  }
+
+  function handleKeyDown(e: React.KeyboardEvent) {
+    if (e.key === 'ArrowDown') {
+      e.preventDefault()
+      setFocusedIndex((prev) => (prev + 1) % filteredOptions.length)
+    }
+
+    if (e.key === 'ArrowUp') {
+      e.preventDefault()
+      setFocusedIndex((prev) => (prev <= 0 ? filteredOptions.length - 1 : prev - 1))
+    }
+
+    if (e.key === 'Enter') {
+      e.preventDefault()
+      if (filteredOptions[focusedIndex]) {
+        handleSelect(filteredOptions[focusedIndex].value)
+      }
+    }
+
+    if (e.key === 'Escape') {
+      e.preventDefault()
+      setOpen(false)
+    }
+  }
+
+  useEffect(() => {
+    if (focusedIndex >= 0 && itemRefs.current[focusedIndex]) {
+      itemRefs.current[focusedIndex].focus()
+    }
+  }, [focusedIndex])
+
+  useEffect(() => {
+    setInternalValue(defaultValue)
+  }, [defaultValue])
+
+  return (
+    <Popover.Root open={open} onOpenChange={setOpen}>
+      <Popover.Trigger asChild>
+        <button
+          className={cn(selectVariants({ size, className }))}
+          type="button"
+          onClick={clearSearch}
+          disabled={disabled}
+          aria-invalid={isInvalid}
+        >
+          {options?.find((opt) => opt.value === internalValue)?.label || placeholder}
+          <ChevronsUpDown className="w-[0.89rem] h-[0.89rem] opacity-50" />
+        </button>
+      </Popover.Trigger>
+
+      <Popover.Portal>
+        <Popover.Content
+          side="bottom"
+          align="start"
+          sideOffset={4}
+          className="bg-neutral-900 w-[var(--radix-popover-trigger-width)] rounded-none hover:rounded-none shadow-lg z-50"
+          onKeyDown={handleKeyDown}
+        >
+          <div className="flex flex-col font-normal text-[0.75rem] leading-[1.25rem]">
+            {isCombobox && (
+              <>
+                <div className="relative">
+                  <Search className="absolute left-3 top-1/2 -translate-y-1/2 w-4 h-4 text-[var(--muted-foreground)]" />
+                  <input
+                    ref={inputRef}
+                    className={cn(
+                      popoverContentItemVariants({ size }),
+                      'bg-neutral-800 placeholder:text-[var(--muted-foreground)] pl-9 focus:ring-0 rounded-none',
+                    )}
+                    placeholder={searchPlaceholder}
+                    onChange={handleSearch}
+                    autoFocus
+                  />
+                </div>
+
+                {!filteredOptions.length && (
+                  <div className={cn(popoverContentItemVariants({ size }), 'rounded-none')}>
+                    {emptyStatePlaceholder}
+                  </div>
+                )}
+              </>
+            )}
+
+            {filteredOptions.map((option, index) => (
+              <Popover.Close asChild key={option.value}>
+                <button
+                  ref={(el) => {
+                    itemRefs.current[index] = el!
+                  }}
+                  tabIndex={-1}
+                  onClick={() => handleSelect(option.value)}
+                  className={cn(popoverContentItemVariants({ size }), 'rounded-none')}
+                >
+                  {option.label}
+                  <Check
+                    className={cn(
+                      'ml-auto w-[1rem] h-[1rem]',
+                      internalValue === option.value ? 'opacity-100' : 'opacity-0',
+                    )}
+                  />
+                </button>
+              </Popover.Close>
+            ))}
+          </div>
+        </Popover.Content>
+      </Popover.Portal>
+    </Popover.Root>
+  )
 }
