+++ packages/ui/src/index.css
@@ -58,6 +58,7 @@
   --color-sidebar-border: var(--sidebar-border);
   --color-sidebar-ring: var(--sidebar-ring);
   --color-neutral-900: #1C1C1E;
+  --color-neutral-950: #141415;
   --color-neutral-gray: #707179;
 }
