+++ registry.json
@@ -73,25 +73,11 @@
     },
      "select": {
       "name": "Select",
-      "dependencies": ["@radix-ui/react-select", "class-variance-authority"],
+      "dependencies": ["@radix-ui/react-popover"],
       "files": [
         {
           "path": "components/ui/select.tsx",
           "contentUrl": "https://raw.githubusercontent.com/roxygens/ui-startkit/refs/heads/main/packages/ui/src/components/ui/select/select.tsx"
-        },
-        {
-          "path": "lib/utils.ts",
-          "contentUrl": "https://raw.githubusercontent.com/roxygens/ui-startkit/refs/heads/main/packages/ui/src/lib/utils.ts"
-        }
-      ]
-    },
-    "combobox": {
-      "name": "Combobox",
-      "dependencies": ["@radix-ui/react-popover"],
-      "files": [
-        {
-          "path": "components/ui/combobox.tsx",
-          "contentUrl": "https://raw.githubusercontent.com/roxygens/ui-startkit/refs/heads/main/packages/ui/src/components/ui/combobox/combobox.tsx"
         },
         {
           "path": "lib/utils.ts",
